fix(filter): передача фильтра между виждетами
осталось
-- применение фильтрации (создание фиьтра при не выбранных категориях)
-- оформление согласно дизайну

+++ places/lib/domain/search_radius.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/cupertino.dart';
 
 class SearchRadius {
-  final int startValue;
-  final int endValue;
+  final double startValue;
+  final double endValue;
 
   SearchRadius({
     @required this.startValue,
